Support across both Twig 2 and 3 for different method signatures
